requirements.txt
python-telegram-bot
requests
bs4
lxml
selenium
webdriver_manager